Complete CRUD operations for user management

+++ src/main/java/com/example/bugtracker/services/UserService.java
@@ -17,16 +17,30 @@ public class UserService {
     // create new user
     public User saveUser(User user) {
         try {
-            return repository.save(user);
+            if (user.getUserName().length() > 8) {
+                throw new ApiRequestException("Username must be only 8 characters");
+            } else {
+                return repository.save(user);
+            }
         } catch (DataIntegrityViolationException ex) {
-            System.out.println("Error");
-            throw  new ApiRequestException("Cannnot creat user");
+            System.out.println("Error" + ex);
+            throw  new ApiRequestException("Username already exists");
         }
     }
 
     // get users
     public User getUserById(int id) {
-        return repository.findById(id).orElse(null);
+        try {
+            User user = repository.findById(id).orElse(null);
+            if (user.equals(null)) {
+                throw new ApiRequestException("There is no user for id " + id);
+            } else {
+                return user;
+            }
+        } catch (NullPointerException ex) {
+            System.out.println("Error " + ex);
+            throw new ApiRequestException("There is no user for id " + id);
+        }
     }
 
     public List<User> getUsers() {
@@ -35,18 +49,36 @@ public class UserService {
 
     // update user
     public User updateUser(User user) {
-        User existingUser = repository.findById(user.getId()).orElse(null);
-        existingUser.setFirstName(user.getFirstName());
-        existingUser.setLastName(user.getLastName());
-        existingUser.setEmail(user.getEmail());
-        existingUser.setPhone(user.getPhone());
-        existingUser.setPassword(user.getPassword());
-        return repository.save(existingUser);
+        try {
+            User existingUser = repository.findById(user.getId()).orElse(null);
+            if (!existingUser.equals(null)) {
+                existingUser.setFirstName(user.getFirstName());
+                existingUser.setLastName(user.getLastName());
+                existingUser.setEmail(user.getEmail());
+                existingUser.setPhone(user.getPhone());
+                existingUser.setPassword(user.getPassword());
+                existingUser.setRole(user.getRole());
+                return repository.save(existingUser);
+            } else {
+                throw new ApiRequestException("There is no user to update for id " + user.getId());
+            }
+        } catch (NullPointerException ex) {
+            throw new ApiRequestException("There is no user to update for id " + user.getId());
+        }
     }
 
     // delete user
     public String deleteUser(int id) {
-        repository.deleteById(id);
-        return "User removed !!";
+        try {
+            User user = repository.findById(id).orElse(null);
+            if (!user.equals(null)) {
+                repository.deleteById(id);
+                return "User removed !!";
+            } else {
+                throw new ApiRequestException("There is no user to delete for id " + id);
+            }
+        } catch (NullPointerException ex) {
+            throw new ApiRequestException("There is no user to delete for id " + id);
+        }
     }
 }
